Add manager for KNoT Datas
Add a class that holds the instances of all knot-datas from a Thing and
offers an interface so that other classes can interact with specific
sensors.

This class works as an abstraction to the management of the Things
sensors, this is specially useful since the operations performed by
the Things generally involve parsing through a complete list of the
sensors which is abstracted by this class.